Changed file structure and added empty tests
